finalize UI and added favorites page

+++ app/src/main/java/com/codev/recruitment/ralphemersonmanzano/navigation/HomeNavigator.kt
@@ -1,11 +1,9 @@
 package com.codev.recruitment.ralphemersonmanzano.navigation
 
-import android.os.Bundle
 import androidx.navigation.NavController
 import com.codev.recruitment.ralphemersonmanzano.NavigationMainDirections
 import com.codev.recruitment.ralphemersonmanzano.details.R
 import com.codev.recruitment.ralphemersonmanzano.home.navigation.HomeNavigation
-import com.codev.recruitment.ralphemersonmanzano.home.navigation.HomeNavigation.Key.KEY_ID
 import dagger.hilt.android.scopes.ActivityScoped
 import javax.inject.Inject
 
@@ -14,7 +12,7 @@ class HomeNavigator @Inject constructor(
     private val navController: NavController
 ): HomeNavigation {
 
-    override fun navigateToDetails(id: Int) {
+    override fun navigateToDetails(id: Long) {
         navController.navigate(NavigationMainDirections.actionToDetails(id, R.id.fragment_details))
     }
